requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
djangorestframework==3.16.0
et_xmlfile==2.0.0
greenlet==3.1.1
joblib==1.4.2
mysql-connector-python==9.3.0
narwhals==1.31.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
plotly==6.0.1
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
